+++ app/src/main/res/layout/fragment_store_detail_available_drinks.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.home.info.StoreDetailAvailableDrinksFragment">
+    tools:context=".ui.store.StoreDetailAvailableDrinksFragment">
 
     <include
         android:id="@+id/toolbar"
