+++ app/src/main/res/layout/fragment_store_detail_info.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.home.info.StoreDetailInfoFragment">
+    tools:context=".ui.store.StoreDetailInfoFragment">
 
     <include
         android:id="@+id/toolbar"
