+++ app/src/main/res/layout/fragment_store_detail_info_main.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:paddingHorizontal="20dp"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.home.info.StoreDetailInfoMainFragment">
+    tools:context=".ui.store.StoreDetailInfoMainFragment">
 
     <include
         android:id="@+id/toolbar"
